refactor(BO): split the system config tabs into specific components
