+++ lib/zaq_web/live/bo/system/system_config_live.ex
@@ -1,23 +1,15 @@
 defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
   use ZaqWeb, :live_view
 
-  import ZaqWeb.Components.SearchableSelect
-
-  alias Phoenix.LiveView.JS
-  alias Zaq.Agent
   alias Zaq.Agent.MCP
-  alias Zaq.Engine.Connect
   alias Zaq.Event
   alias Zaq.NodeRouter
-  alias Zaq.System
   alias Zaq.System.EmbeddingConfig
   alias Zaq.System.ImageToTextConfig
   alias Zaq.System.LLMConfig
   alias Zaq.System.TelemetryConfig
   alias Zaq.Types.EncryptedString
   alias Zaq.Utils.ParseUtils
-  alias ZaqWeb.Components.BOModal
-  alias ZaqWeb.Components.ConnectCredentialForm
 
   def mount(_params, _session, socket) do
     {:ok,
@@ -42,7 +34,8 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
      |> assign(:mcp_page, 1)
      |> assign(:mcp_per_page, 20)
      |> assign(:global_agent_options, global_agent_options())
-     |> assign(:global_default_agent_id, System.get_global_default_agent_id())
+     |> assign(:global_default_agent_id, engine_get_global_default_agent_id())
+     |> assign(:ai_provider_options, provider_options(fn _ -> true end))
      |> assign(:connect_grants_modal, false)
      |> assign(:connect_credential_modal, false)
      |> assign(:connect_credential_action, :edit)
@@ -90,9 +83,9 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
   end
 
   def handle_event("edit_connect_credential", %{"id" => id}, socket) do
-    case Connect.fetch_credential(id) do
+    case engine_connect_fetch_credential(id) do
       {:ok, credential} ->
-        changeset = Connect.change_credential(credential, %{})
+        changeset = engine_connect_change_credential(credential, %{})
 
         {:noreply,
          socket
@@ -121,7 +114,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
 
     changeset =
       credential
-      |> Connect.change_credential(sanitize_connect_credential_params(params))
+      |> engine_connect_change_credential(sanitize_connect_credential_params(params))
       |> Map.put(:action, :validate)
 
     {:noreply,
@@ -134,7 +127,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
   def handle_event("save_connect_credential", %{"credential" => params}, socket) do
     credential = socket.assigns.connect_credential_changeset.data
 
-    case Connect.update_credential(credential, sanitize_connect_credential_params(params)) do
+    case engine_connect_update_credential(credential, sanitize_connect_credential_params(params)) do
       {:ok, _updated} ->
         {:noreply,
          socket
@@ -171,7 +164,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
         {:noreply, put_flash(socket, :error, "Grant not found.")}
 
       grant ->
-        case Connect.delete_grant(grant) do
+        case engine_connect_delete_grant(grant) do
           {:ok, _} ->
             {:noreply,
              socket
@@ -192,7 +185,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
         {:noreply, put_flash(socket, :error, "Grant not found.")}
 
       grant ->
-        case Connect.schedule_refresh(grant) do
+        case engine_connect_schedule_refresh(grant) do
           {:ok, _job} ->
             {:noreply,
              socket
@@ -209,7 +202,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
 
   def handle_event("validate_telemetry", %{"telemetry_config" => params}, socket) do
     changeset =
-      System.get_telemetry_config()
+      engine_get_telemetry_config()
       |> TelemetryConfig.changeset(params)
       |> Map.put(:action, :validate)
 
@@ -217,9 +210,9 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
   end
 
   def handle_event("save_telemetry", %{"telemetry_config" => params}, socket) do
-    changeset = TelemetryConfig.changeset(System.get_telemetry_config(), params)
+    changeset = TelemetryConfig.changeset(engine_get_telemetry_config(), params)
 
-    case System.save_telemetry_config(changeset) do
+    case engine_save_telemetry_config(changeset) do
       {:ok, _} ->
         {:noreply,
          socket
@@ -237,11 +230,11 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
   end
 
   def handle_event("save_global_default_agent", %{"global_default_agent_id" => raw_id}, socket) do
-    case System.set_global_default_agent_id(ParseUtils.parse_optional_int(raw_id)) do
+    case engine_set_global_default_agent_id(ParseUtils.parse_optional_int(raw_id)) do
       :ok ->
         {:noreply,
          socket
-         |> assign(:global_default_agent_id, System.get_global_default_agent_id())
+         |> assign(:global_default_agent_id, engine_get_global_default_agent_id())
          |> put_flash(:info, "Global default agent saved.")}
 
       {:error, reason} ->
@@ -290,7 +283,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
       end
 
     changeset =
-      System.get_llm_config()
+      engine_get_llm_config()
       |> LLMConfig.changeset(params)
       |> Map.put(:action, :validate)
 
@@ -309,9 +302,9 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
   end
 
   def handle_event("save_llm", %{"llm_config" => params}, socket) do
-    changeset = LLMConfig.changeset(System.get_llm_config(), params)
+    changeset = LLMConfig.changeset(engine_get_llm_config(), params)
 
-    case System.save_llm_config(changeset) do
+    case engine_save_llm_config(changeset) do
       {:ok, _} ->
         {:noreply,
          socket
@@ -354,7 +347,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
     params = adjust_embedding_params(params, provider_id, previous_provider, previous_model)
 
     changeset =
-      System.get_embedding_config()
+      engine_get_embedding_config()
       |> EmbeddingConfig.changeset(params)
       |> Map.put(:action, :validate)
 
@@ -402,7 +395,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
     provider_id = provider_from_credential_id(params["credential_id"])
 
     changeset =
-      System.get_image_to_text_config()
+      engine_get_image_to_text_config()
       |> ImageToTextConfig.changeset(params)
       |> Map.put(:action, :validate)
 
@@ -413,9 +406,9 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
   end
 
   def handle_event("save_image_to_text", %{"image_to_text_config" => params}, socket) do
-    changeset = ImageToTextConfig.changeset(System.get_image_to_text_config(), params)
+    changeset = ImageToTextConfig.changeset(engine_get_image_to_text_config(), params)
 
-    case System.save_image_to_text_config(changeset) do
+    case engine_save_image_to_text_config(changeset) do
       {:ok, _} ->
         {:noreply,
          socket
@@ -444,7 +437,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
   end
 
   def handle_event("edit_ai_credential", %{"id" => id}, socket) do
-    credential = System.get_ai_provider_credential!(id)
+    credential = engine_get_ai_provider_credential!(id)
 
     {:noreply,
      socket
@@ -454,7 +447,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
      |> assign(
        :ai_credential_form,
        credential
-       |> System.change_ai_provider_credential(%{})
+       |> engine_change_ai_provider_credential(%{})
        |> to_form(as: :ai_credential)
      )}
   end
@@ -477,8 +470,8 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
   def handle_event("confirm_delete_ai_credential", _params, socket) do
     result =
       socket.assigns.ai_credential_id
-      |> System.get_ai_provider_credential!()
-      |> System.delete_ai_provider_credential()
+      |> engine_get_ai_provider_credential!()
+      |> engine_delete_ai_provider_credential()
 
     case result do
       {:ok, _credential} ->
@@ -516,7 +509,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
     changeset =
       socket.assigns.ai_credential_action
       |> ai_credential_for_action(socket.assigns.ai_credential_id)
-      |> System.change_ai_provider_credential(params)
+      |> engine_change_ai_provider_credential(params)
       |> Map.put(:action, :validate)
 
     {:noreply, assign(socket, :ai_credential_form, to_form(changeset, as: :ai_credential))}
@@ -527,11 +520,11 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
       case socket.assigns.ai_credential_action do
         :edit ->
           socket.assigns.ai_credential_id
-          |> System.get_ai_provider_credential!()
-          |> System.update_ai_provider_credential(params)
+          |> engine_get_ai_provider_credential!()
+          |> engine_update_ai_provider_credential(params)
 
         _ ->
-          System.create_ai_provider_credential(params)
+          engine_create_ai_provider_credential(params)
       end
 
     case result do
@@ -585,7 +578,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
   end
 
   def handle_event("edit_mcp_endpoint", %{"id" => id}, socket) do
-    endpoint = MCP.get_mcp_endpoint!(id)
+    endpoint = agent_get_mcp_endpoint!(id)
 
     {:noreply,
      socket
@@ -593,7 +586,10 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
      |> assign(:mcp_endpoint_id, endpoint.id)
      |> assign(:mcp_endpoint_delete_confirm_modal, false)
      |> assign(:mcp_endpoint_modal, true)
-     |> assign(:mcp_endpoint_form, to_form(MCP.change_mcp_endpoint(endpoint), as: :mcp_endpoint))
+     |> assign(
+       :mcp_endpoint_form,
+       to_form(agent_change_mcp_endpoint(endpoint), as: :mcp_endpoint)
+     )
      |> assign(:mcp_endpoint_rows, mcp_rows(endpoint))}
   end
 
@@ -667,7 +663,8 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
           |> put_flash(:info, "Predefined MCP enabled.")
           |> maybe_put_mcp_runtime_warnings(payload)
 
-        predefined = endpoint.predefined_id && MCP.predefined_catalog()[endpoint.predefined_id]
+        predefined =
+          endpoint.predefined_id && agent_mcp_predefined_catalog()[endpoint.predefined_id]
 
         socket =
           if is_map(predefined) and predefined[:editable] do
@@ -678,7 +675,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
             |> assign(:mcp_endpoint_delete_confirm_modal, false)
             |> assign(
               :mcp_endpoint_form,
-              to_form(MCP.change_mcp_endpoint(endpoint), as: :mcp_endpoint)
+              to_form(agent_change_mcp_endpoint(endpoint), as: :mcp_endpoint)
             )
             |> assign(:mcp_endpoint_rows, mcp_rows(endpoint))
           else
@@ -701,7 +698,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
     changeset =
       socket.assigns.mcp_endpoint_action
       |> mcp_endpoint_for_action(socket.assigns.mcp_endpoint_id)
-      |> MCP.change_mcp_endpoint(parsed)
+      |> agent_change_mcp_endpoint(parsed)
       |> Map.put(:action, :validate)
 
     {:noreply,
@@ -808,9 +805,9 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
   # ── Private ────────────────────────────────────────────────────────────
 
   defp do_save_embedding(socket, params) do
-    changeset = EmbeddingConfig.changeset(System.get_embedding_config(), params)
+    changeset = EmbeddingConfig.changeset(engine_get_embedding_config(), params)
 
-    case System.save_embedding_config(changeset) do
+    case engine_save_embedding_config(changeset) do
       {:ok, _} ->
         {:noreply,
          socket
@@ -829,20 +826,20 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
   end
 
   defp load_telemetry_form(socket) do
-    changeset = TelemetryConfig.changeset(System.get_telemetry_config(), %{})
+    changeset = TelemetryConfig.changeset(engine_get_telemetry_config(), %{})
     assign(socket, :telemetry_form, to_form(changeset, as: :telemetry_config))
   end
 
   defp load_ai_credentials(socket) do
-    assign(socket, :ai_credentials, System.list_ai_provider_credentials())
+    assign(socket, :ai_credentials, engine_list_ai_provider_credentials())
   end
 
   defp load_connect_credentials(socket) do
-    assign(socket, :connect_credentials, Connect.list_credentials())
+    assign(socket, :connect_credentials, engine_connect_list_credentials())
   end
 
   defp load_ai_credential_form(socket) do
-    changeset = System.change_ai_provider_credential(%Zaq.System.AIProviderCredential{}, %{})
+    changeset = engine_change_ai_provider_credential(%Zaq.System.AIProviderCredential{}, %{})
     assign(socket, :ai_credential_form, to_form(changeset, as: :ai_credential))
   end
 
@@ -854,7 +851,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
     }
 
     {entries, total} =
-      MCP.filter_mcp_endpoints(filters,
+      agent_filter_mcp_endpoints(filters,
         page: socket.assigns.mcp_page,
         per_page: socket.assigns.mcp_per_page
       )
@@ -866,7 +863,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
 
   defp load_mcp_endpoint_form(socket) do
     changeset =
-      MCP.change_mcp_endpoint(%MCP.Endpoint{}, %{
+      agent_change_mcp_endpoint(%MCP.Endpoint{}, %{
         "type" => "local",
         "status" => "disabled",
         "timeout_ms" => 5000
@@ -877,7 +874,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
     |> assign(:mcp_endpoint_rows, mcp_rows(%MCP.Endpoint{}))
   end
 
-  defp mcp_endpoint_for_action(:edit, id), do: MCP.get_mcp_endpoint!(id)
+  defp mcp_endpoint_for_action(:edit, id), do: agent_get_mcp_endpoint!(id)
   defp mcp_endpoint_for_action(_, _), do: %MCP.Endpoint{}
 
   defp build_mcp_endpoint_payload(params, rows_state) do
@@ -1183,7 +1180,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
   defp mcp_map_get(_map, _key, default), do: default
 
   defp load_llm_form(socket) do
-    cfg = System.get_llm_config()
+    cfg = engine_get_llm_config()
     provider_id = provider_from_credential_id(cfg.credential_id)
     changeset = LLMConfig.changeset(cfg, %{})
 
@@ -1195,7 +1192,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
   end
 
   defp load_embedding_form(socket) do
-    cfg = System.get_embedding_config()
+    cfg = engine_get_embedding_config()
     provider_id = provider_from_credential_id(cfg.credential_id)
     changeset = EmbeddingConfig.changeset(cfg, %{})
 
@@ -1204,14 +1201,14 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
     |> assign(:embedding_model_options, embedding_model_options(provider_id))
     |> assign(:embedding_form, to_form(changeset, as: :embedding_config))
     |> assign(:embedding_locked, true)
-    |> assign(:embedding_ready, System.embedding_ready?())
+    |> assign(:embedding_ready, engine_embedding_ready?())
     |> assign(:saved_model, cfg.model)
     |> assign(:saved_dimension, cfg.dimension)
     |> assign(:model_changed, false)
   end
 
   defp load_image_to_text_form(socket) do
-    cfg = System.get_image_to_text_config()
+    cfg = engine_get_image_to_text_config()
     provider_id = provider_from_credential_id(cfg.credential_id)
     changeset = ImageToTextConfig.changeset(cfg, %{})
 
@@ -1310,7 +1307,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
   # ── Credential helpers ──────────────────────────────────────────────────
 
   defp credential_options do
-    System.list_ai_provider_credentials()
+    engine_list_ai_provider_credentials()
     |> Enum.map(&{&1.name, Integer.to_string(&1.id)})
   end
 
@@ -1324,7 +1321,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
   end
 
   defp provider_from_credential_id(credential_id) when is_integer(credential_id) do
-    case System.get_ai_provider_credential(credential_id) do
+    case engine_get_ai_provider_credential(credential_id) do
       %{provider: provider} when is_binary(provider) -> provider
       _ -> "custom"
     end
@@ -1332,7 +1329,7 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
 
   defp provider_from_credential_id(_), do: "custom"
 
-  defp ai_credential_for_action(:edit, id), do: System.get_ai_provider_credential!(id)
+  defp ai_credential_for_action(:edit, id), do: engine_get_ai_provider_credential!(id)
   defp ai_credential_for_action(_, _), do: %Zaq.System.AIProviderCredential{}
 
   # ── LLM-specific helpers ───────────────────────────────────────────────
@@ -1432,199 +1429,139 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
     :image in input
   end
 
-  # ── Telemetry Panel ────────────────────────────────────────────────────
-
-  attr :form, :any, required: true
-
-  defp telemetry_panel(assigns) do
-    ~H"""
-    <div class="bg-white rounded-2xl border border-black/[0.06] shadow-sm overflow-hidden">
-      <div class="px-8 py-5 border-b border-black/[0.06] bg-[#fafafa]">
-        <h2 class="font-mono text-[0.95rem] font-bold text-black">Telemetry Collection</h2>
-        <p class="font-mono text-[0.75rem] text-black/40 mt-0.5">
-          Control infra event capture and minimum duration thresholds.
-        </p>
-      </div>
-      <div class="px-8 py-6">
-        <.form
-          id="telemetry-config-form"
-          for={@form}
-          phx-submit="save_telemetry"
-          phx-change="validate_telemetry"
-          class="space-y-5"
-        >
-          <div class="flex items-center justify-between py-2 border-b border-black/[0.05]">
-            <div>
-              <p class="font-mono text-[0.82rem] font-semibold text-black">
-                Capture infra metrics
-              </p>
-              <p class="font-mono text-[0.72rem] text-black/40 mt-0.5">
-                Collect Phoenix request, Repo query, and Oban runtime metrics.
-              </p>
-            </div>
-            <label class="relative inline-flex items-center cursor-pointer">
-              <input type="hidden" name="telemetry_config[capture_infra_metrics]" value="false" />
-              <input
-                type="checkbox"
-                name="telemetry_config[capture_infra_metrics]"
-                value="true"
-                checked={@form[:capture_infra_metrics].value in [true, "true"]}
-                class="sr-only peer"
-              />
-              <div class="w-11 h-6 bg-black/10 peer-checked:bg-[#03b6d4] rounded-full transition-colors after:content-[''] after:absolute after:top-[2px] after:left-[2px] after:bg-white after:rounded-full after:h-5 after:w-5 after:transition-all peer-checked:after:translate-x-5 after:shadow-sm">
-              </div>
-            </label>
-          </div>
-          <div class="grid grid-cols-2 gap-4">
-            <div>
-              <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-                Request Duration Threshold (ms)
-              </label>
-              <input
-                type="number"
-                min="0"
-                name="telemetry_config[request_duration_threshold_ms]"
-                value={@form[:request_duration_threshold_ms].value}
-                phx-debounce="400"
-                placeholder="0"
-                class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa] placeholder:text-black/25 focus:outline-none focus:ring-2 focus:ring-[#03b6d4]/20 focus:border-[#03b6d4] transition-all"
-              />
-              <p
-                :for={{msg, opts} <- @form[:request_duration_threshold_ms].errors}
-                class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-              >
-                {translate_error({msg, opts})}
-              </p>
-            </div>
-            <div>
-              <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-                Repo Query Threshold (ms)
-              </label>
-              <input
-                type="number"
-                min="0"
-                name="telemetry_config[repo_query_duration_threshold_ms]"
-                value={@form[:repo_query_duration_threshold_ms].value}
-                phx-debounce="400"
-                placeholder="0"
-                class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa] placeholder:text-black/25 focus:outline-none focus:ring-2 focus:ring-[#03b6d4]/20 focus:border-[#03b6d4] transition-all"
-              />
-              <p
-                :for={{msg, opts} <- @form[:repo_query_duration_threshold_ms].errors}
-                class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-              >
-                {translate_error({msg, opts})}
-              </p>
-            </div>
-            <div>
-              <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-                No-Answer Alert Threshold (%)
-              </label>
-              <input
-                type="number"
-                min="0"
-                max="100"
-                name="telemetry_config[no_answer_alert_threshold_percent]"
-                value={@form[:no_answer_alert_threshold_percent].value}
-                phx-debounce="400"
-                placeholder="10"
-                class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa] placeholder:text-black/25 focus:outline-none focus:ring-2 focus:ring-[#03b6d4]/20 focus:border-[#03b6d4] transition-all"
-              />
-              <p
-                :for={{msg, opts} <- @form[:no_answer_alert_threshold_percent].errors}
-                class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-              >
-                {translate_error({msg, opts})}
-              </p>
-            </div>
-            <div>
-              <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-                Conversations Response SLA (ms)
-              </label>
-              <input
-                type="number"
-                min="0"
-                name="telemetry_config[conversation_response_sla_ms]"
-                value={@form[:conversation_response_sla_ms].value}
-                phx-debounce="400"
-                placeholder="1500"
-                class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa] placeholder:text-black/25 focus:outline-none focus:ring-2 focus:ring-[#03b6d4]/20 focus:border-[#03b6d4] transition-all"
-              />
-              <p
-                :for={{msg, opts} <- @form[:conversation_response_sla_ms].errors}
-                class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-              >
-                {translate_error({msg, opts})}
-              </p>
-            </div>
-          </div>
-          <div class="bg-[#fafafa] rounded-xl border border-black/5 px-4 py-3">
-            <p class="font-mono text-[0.72rem] text-black/50 leading-relaxed">
-              Thresholds are applied by the telemetry collector and Conversations dashboard alerts.
-              Use <span class="font-semibold text-black/70">0</span> to capture every event.
-            </p>
-          </div>
-          <div class="pt-2">
-            <button
-              type="submit"
-              class="font-mono text-[0.82rem] font-bold px-6 py-3 rounded-xl bg-[#03b6d4] text-white hover:bg-[#029ab3] shadow-sm shadow-[#03b6d4]/20 transition-all"
-            >
-              Save Telemetry Settings
-            </button>
-          </div>
-        </.form>
-      </div>
-    </div>
-    """
+  defp dispatch_engine(action, request \\ %{}) do
+    Event.new(request, :engine, opts: [action: action])
+    |> NodeRouter.dispatch()
+    |> Map.get(:response)
   end
 
-  attr :global_agent_options, :list, required: true
-  attr :global_default_agent_id, :any, required: true
-
-  defp global_panel(assigns) do
-    ~H"""
-    <div class="bg-white rounded-2xl border border-black/[0.06] shadow-sm overflow-hidden">
-      <div class="px-8 py-5 border-b border-black/[0.06] bg-[#fafafa]">
-        <h2 class="font-mono text-[0.95rem] font-bold text-black">Agents</h2>
-        <p class="font-mono text-[0.75rem] text-black/40 mt-0.5">
-          Configure system-wide defaults used when channel-level routing does not select a configured agent.
-        </p>
-      </div>
-      <div class="px-8 py-6">
-        <label class="font-mono text-[0.68rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-          Global Default Agent
-        </label>
-        <form phx-submit="save_global_default_agent" class="flex items-center gap-2">
-          <select
-            id="global-default-agent-select"
-            name="global_default_agent_id"
-            class="w-full max-w-md font-mono text-[0.82rem] text-black border border-black/10 rounded-xl h-10 px-3 bg-white focus:outline-none focus:ring-2 focus:ring-[#03b6d4]/20 focus:border-[#03b6d4]"
-          >
-            <option value="" selected={is_nil(@global_default_agent_id)}>
-              Default Zaq Agent
-            </option>
-            <option
-              :for={{name, id} <- @global_agent_options}
-              value={id}
-              selected={to_string(@global_default_agent_id || "") == to_string(id)}
-            >
-              {name}
-            </option>
-          </select>
-          <button
-            type="submit"
-            class="font-mono text-[0.72rem] px-3 py-2 rounded-lg border border-black/10 text-black/60 hover:text-black hover:border-black/20"
-          >
-            Save
-          </button>
-        </form>
-      </div>
-    </div>
-    """
+  defp dispatch_agent(action, request \\ %{}) do
+    Event.new(request, :agent, opts: [action: action])
+    |> NodeRouter.dispatch()
+    |> Map.get(:response)
   end
+
+  defp unwrap_ok!({:ok, value}), do: value
+  defp unwrap_ok!(value), do: value
+
+  defp engine_get_global_default_agent_id,
+    do: dispatch_engine(:system_config_get_global_default_agent_id)
+
+  defp engine_set_global_default_agent_id(id),
+    do: dispatch_engine(:system_config_set_global_default_agent_id, %{id: id})
+
+  defp engine_get_telemetry_config, do: dispatch_engine(:system_config_get_telemetry_config)
+
+  defp engine_save_telemetry_config(changeset),
+    do: dispatch_engine(:system_config_save_telemetry_config, %{changeset: changeset})
+
+  defp engine_get_llm_config, do: dispatch_engine(:system_config_get_llm_config)
+
+  defp engine_save_llm_config(changeset),
+    do: dispatch_engine(:system_config_save_llm_config, %{changeset: changeset})
+
+  defp engine_get_embedding_config, do: dispatch_engine(:system_config_get_embedding_config)
+
+  defp engine_save_embedding_config(changeset),
+    do: dispatch_engine(:system_config_save_embedding_config, %{changeset: changeset})
+
+  defp engine_embedding_ready?, do: dispatch_engine(:system_config_embedding_ready)
+
+  defp engine_get_image_to_text_config,
+    do: dispatch_engine(:system_config_get_image_to_text_config)
+
+  defp engine_save_image_to_text_config(changeset),
+    do: dispatch_engine(:system_config_save_image_to_text_config, %{changeset: changeset})
+
+  defp engine_list_ai_provider_credentials,
+    do: dispatch_engine(:system_config_list_ai_provider_credentials)
+
+  defp engine_get_ai_provider_credential(id),
+    do: dispatch_engine(:system_config_get_ai_provider_credential, %{id: id})
+
+  defp engine_get_ai_provider_credential!(id) do
+    case dispatch_engine(:system_config_get_ai_provider_credential_bang, %{id: id}) do
+      {:ok, credential} -> credential
+      _ -> raise Ecto.NoResultsError, queryable: Zaq.System.AIProviderCredential
+    end
+  end
+
+  defp engine_change_ai_provider_credential(credential, attrs),
+    do:
+      dispatch_engine(:system_config_change_ai_provider_credential, %{
+        credential: credential,
+        attrs: attrs
+      })
+
+  defp engine_create_ai_provider_credential(attrs),
+    do: dispatch_engine(:system_config_create_ai_provider_credential, %{attrs: attrs})
+
+  defp engine_update_ai_provider_credential(credential, attrs),
+    do:
+      dispatch_engine(:system_config_update_ai_provider_credential, %{
+        credential: credential,
+        attrs: attrs
+      })
+
+  defp engine_delete_ai_provider_credential(credential),
+    do: dispatch_engine(:system_config_delete_ai_provider_credential, %{credential: credential})
+
+  defp engine_connect_list_credentials,
+    do: dispatch_engine(:system_config_connect_list_credentials)
+
+  defp engine_connect_fetch_credential(id),
+    do: dispatch_engine(:connect_fetch_credential, %{credential_id: id})
+
+  defp engine_connect_change_credential(credential, attrs),
+    do:
+      dispatch_engine(:system_config_connect_change_credential, %{
+        credential: credential,
+        attrs: attrs
+      })
+
+  defp engine_connect_update_credential(credential, attrs),
+    do:
+      dispatch_engine(:system_config_connect_update_credential, %{
+        credential: credential,
+        attrs: attrs
+      })
+
+  defp engine_connect_list_grants(credential_id),
+    do: dispatch_engine(:system_config_connect_list_grants, %{credential_id: credential_id})
+
+  defp engine_connect_next_refresh_jobs_for_grants(grants),
+    do: dispatch_engine(:system_config_connect_next_refresh_jobs_for_grants, %{grants: grants})
+
+  defp engine_connect_delete_grant(grant),
+    do: dispatch_engine(:system_config_connect_delete_grant, %{grant: grant})
+
+  defp engine_connect_schedule_refresh(grant),
+    do: dispatch_engine(:system_config_connect_schedule_refresh, %{grant: grant})
+
+  defp agent_get_mcp_endpoint!(id) do
+    dispatch_agent(:system_config_mcp_get_endpoint, %{id: id})
+    |> unwrap_ok!()
+  end
+
+  defp agent_change_mcp_endpoint(endpoint, attrs \\ %{}),
+    do: dispatch_agent(:system_config_mcp_change_endpoint, %{endpoint: endpoint, attrs: attrs})
+
+  defp agent_filter_mcp_endpoints(filters, opts),
+    do:
+      dispatch_agent(:system_config_mcp_filter_endpoints, %{
+        filters: filters,
+        page: opts[:page],
+        per_page: opts[:per_page]
+      })
+
+  defp agent_mcp_predefined_catalog,
+    do: dispatch_agent(:system_config_mcp_predefined_catalog)
+
+  defp agent_list_active_agents,
+    do: dispatch_agent(:system_config_agent_list_active_agents)
 
   defp global_agent_options do
-    Agent.list_active_agents()
+    agent_list_active_agents()
     |> Enum.map(fn agent -> {agent.name, agent.id} end)
   end
 
@@ -1640,1257 +1577,6 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
     )
   end
 
-  # ── LLM Panel ─────────────────────────────────────────────────────────
-
-  attr :form, :any, required: true
-  attr :credential_options, :list, required: true
-  attr :model_options, :list, required: true
-  attr :capabilities, :map, required: true
-
-  defp llm_panel(assigns) do
-    ~H"""
-    <div class="bg-white rounded-2xl border border-black/[0.06] shadow-sm overflow-hidden">
-      <div class="px-8 py-5 border-b border-black/[0.06] bg-[#fafafa]">
-        <h2 class="font-mono text-[0.95rem] font-bold text-black">LLM</h2>
-        <p class="font-mono text-[0.75rem] text-black/40 mt-0.5">
-          OpenAI-compatible language model endpoint used for chat and retrieval.
-        </p>
-      </div>
-      <div class="px-8 py-6">
-        <.form
-          id="llm-config-form"
-          for={@form}
-          phx-submit="save_llm"
-          phx-change="validate_llm"
-          class="space-y-5"
-        >
-          <input type="hidden" name="llm_config[path]" value={@form[:path].value} />
-          <div class="grid grid-cols-2 gap-4">
-            <div>
-              <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-                AI Credential
-              </label>
-              <.searchable_select
-                id="llm-credential-select"
-                name="llm_config[credential_id]"
-                value={to_string(@form[:credential_id].value || "")}
-                options={@credential_options}
-                placeholder="Search credentials..."
-                empty_label="Select a credential..."
-              />
-              <p
-                :for={{msg, opts} <- @form[:credential_id].errors}
-                class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-              >
-                {translate_error({msg, opts})}
-              </p>
-            </div>
-            <div>
-              <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-                Model
-              </label>
-              <.searchable_select
-                :if={@model_options != []}
-                id="llm-model-select"
-                name="llm_config[model]"
-                value={@form[:model].value}
-                options={@model_options}
-                placeholder="Search models..."
-                empty_label="Select a model..."
-              />
-              <input
-                :if={@model_options == []}
-                type="text"
-                name="llm_config[model]"
-                value={@form[:model].value}
-                required
-                phx-debounce="400"
-                placeholder="llama-3.3-70b-instruct"
-                class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa] placeholder:text-black/25 focus:outline-none focus:ring-2 focus:ring-[#03b6d4]/20 focus:border-[#03b6d4] transition-all"
-              />
-              <p
-                :for={{msg, opts} <- @form[:model].errors}
-                class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-              >
-                {translate_error({msg, opts})}
-              </p>
-            </div>
-          </div>
-          <div class="grid grid-cols-2 gap-4">
-            <div>
-              <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-                Temperature
-              </label>
-              <input
-                type="number"
-                min="0"
-                max="2"
-                step="0.1"
-                name="llm_config[temperature]"
-                value={@form[:temperature].value}
-                phx-debounce="400"
-                placeholder="0.0"
-                class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa] placeholder:text-black/25 focus:outline-none focus:ring-2 focus:ring-[#03b6d4]/20 focus:border-[#03b6d4] transition-all"
-              />
-              <p
-                :for={{msg, opts} <- @form[:temperature].errors}
-                class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-              >
-                {translate_error({msg, opts})}
-              </p>
-            </div>
-            <div>
-              <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-                Top-P
-              </label>
-              <input
-                type="number"
-                min="0.01"
-                max="1"
-                step="0.05"
-                name="llm_config[top_p]"
-                value={@form[:top_p].value}
-                phx-debounce="400"
-                placeholder="0.9"
-                class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa] placeholder:text-black/25 focus:outline-none focus:ring-2 focus:ring-[#03b6d4]/20 focus:border-[#03b6d4] transition-all"
-              />
-              <p
-                :for={{msg, opts} <- @form[:top_p].errors}
-                class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-              >
-                {translate_error({msg, opts})}
-              </p>
-            </div>
-          </div>
-          <div class="grid grid-cols-2 gap-4 border-t border-black/[0.06] pt-4">
-            <div>
-              <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-                Max Context Window (tokens)
-              </label>
-              <input
-                type="number"
-                min="1"
-                name="llm_config[max_context_window]"
-                value={@form[:max_context_window].value}
-                phx-debounce="400"
-                placeholder="5000"
-                class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa] placeholder:text-black/25 focus:outline-none focus:ring-2 focus:ring-[#03b6d4]/20 focus:border-[#03b6d4] transition-all"
-              />
-              <p
-                :for={{msg, opts} <- @form[:max_context_window].errors}
-                class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-              >
-                {translate_error({msg, opts})}
-              </p>
-            </div>
-            <div>
-              <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-                Distance Threshold
-              </label>
-              <input
-                type="number"
-                min="0.01"
-                step="0.01"
-                name="llm_config[distance_threshold]"
-                value={@form[:distance_threshold].value}
-                phx-debounce="400"
-                placeholder="1.2"
-                class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa] placeholder:text-black/25 focus:outline-none focus:ring-2 focus:ring-[#03b6d4]/20 focus:border-[#03b6d4] transition-all"
-              />
-              <p
-                :for={{msg, opts} <- @form[:distance_threshold].errors}
-                class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-              >
-                {translate_error({msg, opts})}
-              </p>
-            </div>
-          </div>
-          <details
-            id="llm-fusion-advanced"
-            phx-hook="DetailsKeepOpen"
-            class="border border-black/[0.08] rounded-xl overflow-hidden"
-          >
-            <summary class="cursor-pointer select-none px-4 py-3 font-mono text-[0.75rem] font-semibold text-black/50 uppercase tracking-wider hover:bg-black/[0.02] transition-colors list-none flex items-center gap-2">
-              <svg
-                class="w-3.5 h-3.5 transition-transform details-open:rotate-90"
-                viewBox="0 0 20 20"
-                fill="currentColor"
-              >
-                <path
-                  fill-rule="evenodd"
-                  d="M7.21 14.77a.75.75 0 0 1 .02-1.06L11.168 10 7.23 6.29a.75.75 0 1 1 1.04-1.08l4.5 4.25a.75.75 0 0 1 0 1.08l-4.5 4.25a.75.75 0 0 1-1.06-.02Z"
-                  clip-rule="evenodd"
-                />
-              </svg>
-              Advanced — Hybrid Search Fusion Weights
-            </summary>
-            <div class="px-4 pb-4 pt-3 bg-black/[0.01] border-t border-black/[0.06] grid grid-cols-2 gap-4">
-              <div>
-                <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-                  BM25 Weight
-                </label>
-                <input
-                  type="number"
-                  min="0"
-                  max="1"
-                  step="0.01"
-                  name="llm_config[fusion_bm25_weight]"
-                  value={@form[:fusion_bm25_weight].value}
-                  phx-debounce="400"
-                  placeholder="0.5"
-                  class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa] placeholder:text-black/25 focus:outline-none focus:ring-2 focus:ring-[#03b6d4]/20 focus:border-[#03b6d4] transition-all"
-                />
-                <p
-                  :for={{msg, opts} <- @form[:fusion_bm25_weight].errors}
-                  class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-                >
-                  {translate_error({msg, opts})}
-                </p>
-              </div>
-              <div>
-                <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-                  Vector Weight
-                </label>
-                <input
-                  type="number"
-                  min="0"
-                  max="1"
-                  step="0.01"
-                  name="llm_config[fusion_vector_weight]"
-                  value={@form[:fusion_vector_weight].value}
-                  phx-debounce="400"
-                  placeholder="0.5"
-                  class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa] placeholder:text-black/25 focus:outline-none focus:ring-2 focus:ring-[#03b6d4]/20 focus:border-[#03b6d4] transition-all"
-                />
-                <p
-                  :for={{msg, opts} <- @form[:fusion_vector_weight].errors}
-                  class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-                >
-                  {translate_error({msg, opts})}
-                </p>
-              </div>
-            </div>
-          </details>
-          <div class="pt-2">
-            <button
-              type="submit"
-              class="font-mono text-[0.82rem] font-bold px-6 py-3 rounded-xl bg-[#03b6d4] text-white hover:bg-[#029ab3] shadow-sm shadow-[#03b6d4]/20 transition-all"
-            >
-              Save LLM Settings
-            </button>
-          </div>
-        </.form>
-      </div>
-    </div>
-    """
-  end
-
-  # ── Embedding Panel ───────────────────────────────────────────────────
-
-  attr :form, :any, required: true
-  attr :credential_options, :list, required: true
-  attr :model_options, :list, required: true
-  attr :locked, :boolean, default: false
-  attr :unlock_modal, :boolean, default: false
-  attr :model_changed, :boolean, default: false
-  attr :save_confirm_modal, :boolean, default: false
-
-  defp embedding_panel(assigns) do
-    ~H"""
-    <%!-- Unlock model — informational note --%>
-    <div
-      :if={@unlock_modal}
-      class="fixed inset-0 z-50 flex items-center justify-center bg-black/30 backdrop-blur-sm"
-    >
-      <div class="bg-white rounded-2xl shadow-xl border border-black/[0.08] w-full max-w-md mx-4 p-6">
-        <h3 class="font-mono text-[0.95rem] font-bold text-black mb-2">Unlock Model Selection</h3>
-        <p class="font-mono text-[0.8rem] text-black/60 leading-relaxed mb-5">
-          You are about to unlock model selection. If you pick a different model, saving will permanently delete all existing embeddings and require full re-ingestion.
-        </p>
-        <div class="flex justify-end gap-3">
-          <button
-            type="button"
-            phx-click="cancel_unlock_embedding"
-            class="font-mono text-[0.82rem] px-5 py-2.5 rounded-xl border border-black/10 text-black/60 hover:bg-black/[0.04] transition-all"
-          >
-            Cancel
-          </button>
-          <button
-            type="button"
-            phx-click="confirm_unlock_embedding"
-            class="font-mono text-[0.82rem] font-bold px-5 py-2.5 rounded-xl bg-[#03b6d4] text-white hover:bg-[#029ab3] transition-all"
-          >
-            Unlock
-          </button>
-        </div>
-      </div>
-    </div>
-
-    <%!-- Destructive save confirmation modal --%>
-    <div
-      :if={@save_confirm_modal}
-      class="fixed inset-0 z-50 flex items-center justify-center bg-black/30 backdrop-blur-sm"
-    >
-      <div class="bg-white rounded-2xl shadow-xl border border-black/[0.08] w-full max-w-md mx-4 p-6">
-        <h3 class="font-mono text-[0.95rem] font-bold text-black mb-2">Delete All Embeddings?</h3>
-        <p class="font-mono text-[0.8rem] text-black/60 leading-relaxed mb-5">
-          All existing embeddings will be permanently deleted and full re-ingestion will be required. This cannot be undone.
-        </p>
-        <div class="flex justify-end gap-3">
-          <button
-            type="button"
-            phx-click="cancel_save_embedding"
-            class="font-mono text-[0.82rem] px-5 py-2.5 rounded-xl border border-black/10 text-black/60 hover:bg-black/[0.04] transition-all"
-          >
-            Cancel
-          </button>
-          <button
-            type="button"
-            phx-click="confirm_save_embedding"
-            class="font-mono text-[0.82rem] font-bold px-5 py-2.5 rounded-xl bg-red-500 text-white hover:bg-red-600 transition-all"
-          >
-            Proceed
-          </button>
-        </div>
-      </div>
-    </div>
-
-    <div class="bg-white rounded-2xl border border-black/[0.06] shadow-sm overflow-hidden">
-      <div class="px-8 py-5 border-b border-black/[0.06] bg-[#fafafa]">
-        <h2 class="font-mono text-[0.95rem] font-bold text-black">Embedding</h2>
-        <p class="font-mono text-[0.75rem] text-black/40 mt-0.5">
-          OpenAI-compatible embedding endpoint used for vector search.
-        </p>
-      </div>
-      <div class="px-8 py-6">
-        <.form
-          id="embedding-config-form"
-          for={@form}
-          phx-submit="save_embedding"
-          phx-change="validate_embedding"
-          class="space-y-5"
-        >
-          <div class="grid grid-cols-2 gap-4 items-start">
-            <div>
-              <div class="h-7 flex items-center mb-2">
-                <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider">
-                  AI Credential
-                </label>
-              </div>
-              <.searchable_select
-                id="embedding-credential-select"
-                name="embedding_config[credential_id]"
-                value={to_string(@form[:credential_id].value || "")}
-                options={@credential_options}
-                placeholder="Search credentials..."
-                empty_label="Select a credential..."
-              />
-              <p
-                :for={{msg, opts} <- @form[:credential_id].errors}
-                class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-              >
-                {translate_error({msg, opts})}
-              </p>
-            </div>
-            <div>
-              <div class="h-7 flex items-center justify-between mb-2">
-                <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider">
-                  Model
-                </label>
-                <div :if={@locked} class="flex items-center gap-2">
-                  <span class="flex items-center gap-1 font-mono text-[0.68rem] font-semibold text-emerald-600 bg-emerald-50 border border-emerald-200 px-2 py-1 rounded-md">
-                    <svg class="w-3 h-3" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                      <path
-                        stroke-linecap="round"
-                        stroke-linejoin="round"
-                        stroke-width="2"
-                        d="M12 15v2m-6 4h12a2 2 0 002-2v-6a2 2 0 00-2-2H6a2 2 0 00-2 2v6a2 2 0 002 2zm10-10V7a4 4 0 00-8 0v4h8z"
-                      />
-                    </svg>
-                    Locked
-                  </span>
-                  <button
-                    type="button"
-                    phx-click="unlock_embedding"
-                    class="font-mono text-[0.68rem] font-semibold px-2 py-1 rounded-md border border-amber-300 text-amber-600 bg-amber-50 hover:bg-amber-100 transition-all"
-                  >
-                    Unlock
-                  </button>
-                </div>
-              </div>
-              <div class={[@locked && "opacity-50 pointer-events-none"]}>
-                <.searchable_select
-                  :if={@model_options != []}
-                  id="embedding-model-select"
-                  name="embedding_config[model]"
-                  value={@form[:model].value}
-                  options={@model_options}
-                  placeholder="Search models..."
-                  empty_label="Select a model..."
-                />
-                <input
-                  :if={@model_options == []}
-                  type="text"
-                  name="embedding_config[model]"
-                  value={@form[:model].value}
-                  phx-debounce="400"
-                  placeholder="bge-multilingual-gemma2"
-                  disabled={@locked}
-                  class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa] placeholder:text-black/25 focus:outline-none focus:ring-2 focus:ring-[#03b6d4]/20 focus:border-[#03b6d4] transition-all"
-                />
-              </div>
-              <p
-                :for={{msg, opts} <- @form[:model].errors}
-                class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-              >
-                {translate_error({msg, opts})}
-              </p>
-            </div>
-          </div>
-          <div>
-            <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-              Dimension
-            </label>
-            <input
-              type="number"
-              min="1"
-              name="embedding_config[dimension]"
-              value={@form[:dimension].value}
-              phx-debounce="400"
-              placeholder="3584"
-              disabled={@locked}
-              class={[
-                "w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 placeholder:text-black/25 transition-all",
-                if(@locked,
-                  do: "bg-black/[0.03] text-black/40 cursor-not-allowed",
-                  else:
-                    "bg-[#fafafa] focus:outline-none focus:ring-2 focus:ring-[#03b6d4]/20 focus:border-[#03b6d4]"
-                )
-              ]}
-            />
-            <p
-              :for={{msg, opts} <- @form[:dimension].errors}
-              class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-            >
-              {translate_error({msg, opts})}
-            </p>
-            <p
-              :if={not @locked and @form[:dimension].value not in [nil, ""]}
-              class="font-mono text-[0.72rem] text-amber-600 mt-1.5"
-            >
-              Changing the model or dimension will permanently delete all chunks and require re-ingestion.
-            </p>
-          </div>
-          <div class="grid grid-cols-2 gap-4 border-t border-black/[0.06] pt-4">
-            <div>
-              <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-                Chunk Min Tokens
-              </label>
-              <input
-                type="number"
-                min="1"
-                name="embedding_config[chunk_min_tokens]"
-                value={@form[:chunk_min_tokens].value}
-                phx-debounce="400"
-                placeholder="400"
-                class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa] placeholder:text-black/25 focus:outline-none focus:ring-2 focus:ring-[#03b6d4]/20 focus:border-[#03b6d4] transition-all"
-              />
-              <p
-                :for={{msg, opts} <- @form[:chunk_min_tokens].errors}
-                class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-              >
-                {translate_error({msg, opts})}
-              </p>
-            </div>
-            <div>
-              <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-                Chunk Max Tokens
-              </label>
-              <input
-                type="number"
-                min="1"
-                name="embedding_config[chunk_max_tokens]"
-                value={@form[:chunk_max_tokens].value}
-                phx-debounce="400"
-                placeholder="900"
-                class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa] placeholder:text-black/25 focus:outline-none focus:ring-2 focus:ring-[#03b6d4]/20 focus:border-[#03b6d4] transition-all"
-              />
-              <p
-                :for={{msg, opts} <- @form[:chunk_max_tokens].errors}
-                class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-              >
-                {translate_error({msg, opts})}
-              </p>
-            </div>
-          </div>
-          <div class="pt-2">
-            <button
-              type="submit"
-              class={[
-                "font-mono text-[0.82rem] font-bold px-6 py-3 rounded-xl text-white shadow-sm transition-all",
-                if(@model_changed,
-                  do: "bg-red-500 hover:bg-red-600 shadow-red-500/20",
-                  else: "bg-[#03b6d4] hover:bg-[#029ab3] shadow-[#03b6d4]/20"
-                )
-              ]}
-            >
-              Save Embedding Settings
-            </button>
-          </div>
-        </.form>
-      </div>
-    </div>
-    """
-  end
-
-  # ── Image to Text Panel ───────────────────────────────────────────────
-
-  attr :form, :any, required: true
-  attr :credential_options, :list, required: true
-  attr :model_options, :list, required: true
-
-  defp image_to_text_panel(assigns) do
-    ~H"""
-    <div class="bg-white rounded-2xl border border-black/[0.06] shadow-sm">
-      <div class="px-8 py-5 border-b border-black/[0.06] bg-[#fafafa] rounded-t-2xl">
-        <h2 class="font-mono text-[0.95rem] font-bold text-black">Image to Text</h2>
-        <p class="font-mono text-[0.75rem] text-black/40 mt-0.5">
-          Vision model endpoint used to extract text from images and PDFs during ingestion.
-        </p>
-      </div>
-      <div class="px-8 py-6">
-        <.form
-          id="image-to-text-config-form"
-          for={@form}
-          phx-submit="save_image_to_text"
-          phx-change="validate_image_to_text"
-          class="space-y-5"
-        >
-          <div class="grid grid-cols-2 gap-4">
-            <div>
-              <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-                AI Credential
-              </label>
-              <.searchable_select
-                id="image-to-text-credential-select"
-                name="image_to_text_config[credential_id]"
-                value={to_string(@form[:credential_id].value || "")}
-                options={@credential_options}
-                placeholder="Search credentials..."
-                empty_label="Select a credential..."
-              />
-              <p
-                :for={{msg, opts} <- @form[:credential_id].errors}
-                class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-              >
-                {translate_error({msg, opts})}
-              </p>
-            </div>
-            <div>
-              <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-                Model
-              </label>
-              <.searchable_select
-                :if={@model_options != []}
-                id="image-to-text-model-select"
-                name="image_to_text_config[model]"
-                value={@form[:model].value}
-                options={@model_options}
-                placeholder="Search models..."
-                empty_label="Select a model..."
-              />
-              <input
-                :if={@model_options == []}
-                type="text"
-                name="image_to_text_config[model]"
-                value={@form[:model].value}
-                phx-debounce="400"
-                placeholder="pixtral-12b-2409"
-                class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa] placeholder:text-black/25 focus:outline-none focus:ring-2 focus:ring-[#03b6d4]/20 focus:border-[#03b6d4] transition-all"
-              />
-              <p
-                :for={{msg, opts} <- @form[:model].errors}
-                class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-              >
-                {translate_error({msg, opts})}
-              </p>
-            </div>
-          </div>
-          <div class="pt-2">
-            <button
-              type="submit"
-              class="font-mono text-[0.82rem] font-bold px-6 py-3 rounded-xl bg-[#03b6d4] text-white hover:bg-[#029ab3] shadow-sm shadow-[#03b6d4]/20 transition-all"
-            >
-              Save Image to Text Settings
-            </button>
-          </div>
-        </.form>
-      </div>
-    </div>
-    """
-  end
-
-  # ── MCP Administration Panel ────────────────────────────────────────────
-
-  attr :entries, :list, required: true
-  attr :total_count, :integer, required: true
-  attr :page, :integer, required: true
-  attr :per_page, :integer, required: true
-  attr :filter_name, :string, required: true
-  attr :filter_type, :string, required: true
-  attr :filter_status, :string, required: true
-  attr :form, :any, required: true
-  attr :rows, :map, required: true
-  attr :modal, :boolean, required: true
-  attr :action, :atom, required: true
-  attr :delete_confirm_modal, :boolean, required: true
-
-  defp mcp_panel(assigns) do
-    ~H"""
-    <div class="bg-white rounded-2xl border border-black/[0.06] shadow-sm overflow-hidden">
-      <div class="px-8 py-5 border-b border-black/[0.06] bg-[#fafafa] flex items-center justify-between">
-        <div>
-          <h2 class="font-mono text-[0.95rem] font-bold text-black">MCP Administration</h2>
-          <p class="font-mono text-[0.75rem] text-black/40 mt-0.5">
-            Manage Model Context Protocol endpoints and test tool loading.
-          </p>
-        </div>
-        <button
-          type="button"
-          phx-click="new_mcp_endpoint"
-          class="font-mono text-[0.75rem] font-bold px-4 py-2 rounded-lg bg-[#03b6d4] text-white hover:bg-[#029ab3] transition-all"
-        >
-          + Add MCP
-        </button>
-      </div>
-
-      <form
-        phx-change="filter_mcp_endpoints"
-        class="px-8 py-4 border-b border-black/[0.06] grid grid-cols-3 gap-3"
-      >
-        <input
-          type="text"
-          name="mcp_filter_name"
-          value={@filter_name}
-          phx-debounce="300"
-          placeholder="Search by name"
-          class="font-mono text-[0.82rem] text-black border border-black/10 rounded-lg h-10 px-3 bg-[#fafafa]"
-        />
-        <select
-          name="mcp_filter_type"
-          class="font-mono text-[0.82rem] text-black border border-black/10 rounded-lg h-10 px-3 bg-[#fafafa]"
-        >
-          <option value="all" selected={@filter_type == "all"}>All types</option>
-          <option value="local" selected={@filter_type == "local"}>Local</option>
-          <option value="remote" selected={@filter_type == "remote"}>Remote</option>
-        </select>
-        <select
-          name="mcp_filter_status"
-          class="font-mono text-[0.82rem] text-black border border-black/10 rounded-lg h-10 px-3 bg-[#fafafa]"
-        >
-          <option value="all" selected={@filter_status == "all"}>All statuses</option>
-          <option value="enabled" selected={@filter_status == "enabled"}>Enabled</option>
-          <option value="disabled" selected={@filter_status == "disabled"}>Disabled</option>
-        </select>
-      </form>
-
-      <div :if={@entries == []} class="px-8 py-10 text-center">
-        <p class="font-mono text-[0.85rem] text-black/50">No MCP entries found.</p>
-      </div>
-
-      <div :if={@entries != []} class="divide-y divide-black/[0.06]">
-        <div :for={entry <- @entries} class="px-8 py-4 flex items-center justify-between gap-4">
-          <div class="min-w-0 flex items-start gap-3">
-            <div class="mt-0.5 inline-flex h-7 w-7 shrink-0 items-center justify-center overflow-hidden rounded-lg">
-              <ZaqWeb.Components.MCPEndpointIcons.icon
-                endpoint_key={entry.predefined_id}
-                class="h-7 w-7"
-              />
-            </div>
-            <div class="min-w-0">
-              <div class="flex items-center gap-2">
-                <p class="font-mono text-[0.82rem] font-semibold text-black truncate">{entry.name}</p>
-                <span class={[
-                  "h-2 w-2 rounded-full",
-                  if(entry.status == "enabled", do: "bg-emerald-500", else: "bg-red-500")
-                ]} />
-              </div>
-              <p class="font-mono text-[0.7rem] text-black/45 mt-0.5">
-                {entry.type}<span :if={entry.predefined?}> · predefined</span>
-              </p>
-              <p :if={entry.description} class="font-mono text-[0.7rem] text-black/35 mt-0.5 truncate">
-                {entry.description}
-              </p>
-            </div>
-          </div>
-          <div class="flex items-center gap-2">
-            <button
-              :if={not entry.persisted? and entry.predefined?}
-              type="button"
-              phx-click="enable_predefined_mcp"
-              phx-value-predefined_id={entry.predefined_id}
-              class="font-mono text-[0.72rem] px-3 py-1.5 rounded-lg border border-black/10 text-black/70 hover:bg-black/[0.04]"
-            >
-              Enable
-              <span :if={entry.auto_enabled} class="ml-1" title="Immediately available">⚡</span>
-            </button>
-            <button
-              :if={entry.persisted?}
-              type="button"
-              phx-click="edit_mcp_endpoint"
-              phx-value-id={entry.id}
-              class="font-mono text-[0.72rem] px-3 py-1.5 rounded-lg border border-black/10 text-black/70 hover:bg-black/[0.04]"
-            >
-              Edit
-            </button>
-            <.loading_action_button
-              :if={entry.persisted?}
-              id={"mcp-test-button-#{entry.id}"}
-              phx-click="test_mcp_endpoint"
-              phx-value-id={entry.id}
-              label="Test"
-              loading_label="Testing..."
-              class="mcp-test-button font-mono text-[0.72rem] px-3 py-1.5 rounded-lg bg-[#03b6d4] text-white hover:bg-[#029ab3]"
-            />
-          </div>
-        </div>
-      </div>
-
-      <div
-        :if={@total_count > 0}
-        class="px-8 py-3 border-t border-black/[0.06] flex items-center justify-between"
-      >
-        <span class="font-mono text-[0.68rem] text-black/40">
-          {@page * @per_page - @per_page + 1}–{min(@page * @per_page, @total_count)} of {@total_count}
-        </span>
-        <div class="flex gap-1">
-          <button
-            :if={@page > 1}
-            type="button"
-            phx-click="change_mcp_page"
-            phx-value-page={@page - 1}
-            class="font-mono text-[0.7rem] px-3 py-1.5 rounded-lg border border-black/12 text-black/60 hover:bg-black/5"
-          >
-            ← Prev
-          </button>
-          <button
-            :if={@page * @per_page < @total_count}
-            type="button"
-            phx-click="change_mcp_page"
-            phx-value-page={@page + 1}
-            class="font-mono text-[0.7rem] px-3 py-1.5 rounded-lg border border-black/12 text-black/60 hover:bg-black/5"
-          >
-            Next →
-          </button>
-        </div>
-      </div>
-    </div>
-
-    <BOModal.form_dialog
-      :if={@modal}
-      id="mcp-endpoint-modal"
-      cancel_event="close_mcp_endpoint_modal"
-      title={if @action == :edit, do: "Edit MCP Endpoint", else: "New MCP Endpoint"}
-      max_width_class="max-w-3xl"
-    >
-      <.form
-        id="mcp-endpoint-form"
-        for={@form}
-        phx-change="validate_mcp_endpoint"
-        phx-submit="save_mcp_endpoint"
-        class="space-y-4"
-      >
-        <p
-          :for={{msg, opts} <- Keyword.get_values(@form.errors, :base)}
-          class="whitespace-pre-line font-mono text-[0.72rem] text-red-500 bg-red-50 border border-red-100 rounded-xl px-3 py-2"
-        >
-          {translate_error({msg, opts})}
-        </p>
-
-        <div class="grid grid-cols-1 gap-4 md:grid-cols-[minmax(0,1fr)_160px_160px]">
-          <div>
-            <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-              Name
-            </label>
-            <input
-              type="text"
-              name="mcp_endpoint[name]"
-              value={@form[:name].value}
-              class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa]"
-            />
-          </div>
-          <div>
-            <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-              Status
-            </label>
-            <label class="h-11 px-1 inline-flex items-center gap-3 cursor-pointer">
-              <input
-                type="hidden"
-                name="mcp_endpoint[status]"
-                value={if @form[:status].value == "enabled", do: "enabled", else: "disabled"}
-              />
-              <input
-                type="checkbox"
-                checked={@form[:status].value == "enabled"}
-                class="sr-only peer"
-                phx-click={
-                  if @form[:status].value == "enabled" do
-                    JS.set_attribute({"value", "disabled"},
-                      to: "#mcp-endpoint-form input[name='mcp_endpoint[status]']"
-                    )
-                  else
-                    JS.set_attribute({"value", "enabled"},
-                      to: "#mcp-endpoint-form input[name='mcp_endpoint[status]']"
-                    )
-                  end
-                  |> JS.dispatch("change", to: "#mcp-endpoint-form")
-                }
-              />
-              <div class="w-11 h-6 bg-black/10 peer-checked:bg-[#03b6d4] rounded-full transition-colors after:content-[''] after:absolute after:top-[2px] after:left-[2px] after:bg-white after:rounded-full after:h-5 after:w-5 after:transition-all peer-checked:after:translate-x-5 after:shadow-sm relative">
-              </div>
-              <span class="font-mono text-[0.78rem] text-black/70">
-                {if @form[:status].value == "enabled", do: "Enabled", else: "Disabled"}
-              </span>
-            </label>
-          </div>
-          <div>
-            <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-              Timeout (ms)
-            </label>
-            <input
-              type="number"
-              min="1"
-              name="mcp_endpoint[timeout_ms]"
-              value={@form[:timeout_ms].value || 5000}
-              class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa]"
-            />
-          </div>
-        </div>
-
-        <input
-          type="hidden"
-          name="mcp_endpoint[predefined_id]"
-          value={@form[:predefined_id].value || ""}
-        />
-
-        <div class="grid grid-cols-1 gap-4 md:grid-cols-[160px_minmax(0,1fr)]">
-          <div>
-            <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-              Type
-            </label>
-            <select
-              name="mcp_endpoint[type]"
-              class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa]"
-            >
-              <option value="local" selected={@form[:type].value == "local"}>local</option>
-              <option value="remote" selected={@form[:type].value == "remote"}>remote</option>
-            </select>
-          </div>
-
-          <div :if={@form[:type].value == "local"}>
-            <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-              Command
-            </label>
-            <input
-              type="text"
-              name="mcp_endpoint[command]"
-              value={@form[:command].value || ""}
-              class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa]"
-            />
-          </div>
-
-          <div :if={@form[:type].value == "remote"}>
-            <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-              URL
-            </label>
-            <input
-              type="text"
-              name="mcp_endpoint[url]"
-              value={@form[:url].value || ""}
-              class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa]"
-            />
-          </div>
-        </div>
-
-        <div :if={@form[:type].value == "local"} class="space-y-4">
-          <.mcp_args_rows rows={@rows.args} collection="args" label="Args" placeholder="--flag" />
-
-          <.mcp_kv_rows
-            rows={@rows.environments}
-            collection="environments"
-            label="Environments"
-            key_placeholder="KEY"
-            value_placeholder="value"
-          />
-
-          <.mcp_secret_kv_rows
-            rows={@rows.secret_environments}
-            collection="secret_environments"
-            label="Secret environments"
-            key_placeholder="SECRET_KEY"
-          />
-        </div>
-
-        <div :if={@form[:type].value == "remote"} class="space-y-4">
-          <.mcp_kv_rows
-            rows={@rows.headers}
-            collection="headers"
-            label="Headers"
-            key_placeholder="Header-Name"
-            value_placeholder="value"
-          />
-
-          <.mcp_secret_kv_rows
-            rows={@rows.secret_headers}
-            collection="secret_headers"
-            label="Secret headers"
-            key_placeholder="Authorization"
-          />
-        </div>
-
-        <div>
-          <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-            Settings (JSON)
-          </label>
-          <textarea
-            name="mcp_endpoint[settings_text]"
-            rows="4"
-            class="w-full font-mono text-[0.84rem] text-black border border-black/10 rounded-xl px-4 py-3 bg-[#fafafa]"
-          >{@rows.settings}</textarea>
-        </div>
-      </.form>
-
-      <:actions>
-        <button
-          :if={@action == :edit}
-          type="button"
-          phx-click="open_delete_mcp_endpoint_confirm"
-          class="inline-flex items-center gap-2 font-mono text-[0.8rem] px-4 py-2 rounded-lg border border-red-200 text-red-600 hover:bg-red-50"
-        >
-          <.icon name="hero-trash" class="h-4 w-4" /> Delete endpoint
-        </button>
-        <button
-          type="submit"
-          form="mcp-endpoint-form"
-          class="font-mono text-[0.8rem] font-bold px-4 py-2 rounded-lg bg-[#03b6d4] text-white hover:bg-[#029ab3]"
-        >
-          Save endpoint
-        </button>
-      </:actions>
-
-      <ZaqWeb.Components.BOModal.confirm_dialog
-        :if={@delete_confirm_modal}
-        id="mcp-endpoint-delete-confirm"
-        cancel_event="cancel_delete_mcp_endpoint"
-        confirm_event="confirm_delete_mcp_endpoint"
-        title="Delete MCP Endpoint?"
-        message="This action removes the endpoint. Associated runtime tools will be unsynced from active agents."
-        confirm_label="Delete"
-        cancel_label="Cancel"
-      />
-    </BOModal.form_dialog>
-    """
-  end
-
-  attr :rows, :list, required: true
-  attr :collection, :string, required: true
-  attr :label, :string, required: true
-  attr :placeholder, :string, default: ""
-
-  defp mcp_args_rows(assigns) do
-    ~H"""
-    <div>
-      <div class="mb-2 flex items-center justify-between">
-        <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider">
-          {@label}
-        </label>
-        <button
-          type="button"
-          phx-click="add_mcp_row"
-          phx-value-collection={@collection}
-          class="font-mono text-[0.68rem] px-2 py-1 rounded border border-black/10 text-black/60 hover:bg-black/[0.04]"
-        >
-          + Add
-        </button>
-      </div>
-      <div class="space-y-2">
-        <div
-          :for={{row, idx} <- Enum.with_index(@rows)}
-          class="grid grid-cols-[minmax(0,1fr)_auto] items-center gap-2"
-        >
-          <input
-            type="text"
-            name={"mcp_endpoint[#{@collection}_rows][#{idx}][value]"}
-            value={row["value"]}
-            placeholder={@placeholder}
-            class="w-full font-mono text-[0.82rem] text-black border border-black/10 rounded-lg h-10 px-3 bg-[#fafafa]"
-          />
-          <.mcp_row_delete_button collection={@collection} index={idx} />
-        </div>
-      </div>
-    </div>
-    """
-  end
-
-  attr :rows, :list, required: true
-  attr :collection, :string, required: true
-  attr :label, :string, required: true
-  attr :key_placeholder, :string, default: "KEY"
-  attr :value_placeholder, :string, default: "value"
-
-  defp mcp_kv_rows(assigns) do
-    ~H"""
-    <div>
-      <div class="mb-2 flex items-center justify-between">
-        <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider">
-          {@label}
-        </label>
-        <button
-          type="button"
-          phx-click="add_mcp_row"
-          phx-value-collection={@collection}
-          class="font-mono text-[0.68rem] px-2 py-1 rounded border border-black/10 text-black/60 hover:bg-black/[0.04]"
-        >
-          + Add
-        </button>
-      </div>
-      <div class="space-y-2">
-        <div
-          :for={{row, idx} <- Enum.with_index(@rows)}
-          class="grid grid-cols-[minmax(0,1fr)_minmax(0,1fr)_auto] items-center gap-2"
-        >
-          <input
-            type="text"
-            name={"mcp_endpoint[#{@collection}_rows][#{idx}][key]"}
-            value={row["key"]}
-            placeholder={@key_placeholder}
-            class="w-full font-mono text-[0.82rem] text-black border border-black/10 rounded-lg h-10 px-3 bg-[#fafafa]"
-          />
-          <input
-            type="text"
-            name={"mcp_endpoint[#{@collection}_rows][#{idx}][value]"}
-            value={row["value"]}
-            placeholder={@value_placeholder}
-            class="w-full font-mono text-[0.82rem] text-black border border-black/10 rounded-lg h-10 px-3 bg-[#fafafa]"
-          />
-          <.mcp_row_delete_button collection={@collection} index={idx} />
-        </div>
-      </div>
-    </div>
-    """
-  end
-
-  attr :rows, :list, required: true
-  attr :collection, :string, required: true
-  attr :label, :string, required: true
-  attr :key_placeholder, :string, default: "KEY"
-
-  defp mcp_secret_kv_rows(assigns) do
-    ~H"""
-    <div>
-      <div class="mb-2 flex items-center justify-between">
-        <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider">
-          {@label}
-        </label>
-        <button
-          type="button"
-          phx-click="add_mcp_row"
-          phx-value-collection={@collection}
-          class="font-mono text-[0.68rem] px-2 py-1 rounded border border-black/10 text-black/60 hover:bg-black/[0.04]"
-        >
-          + Add
-        </button>
-      </div>
-      <div class="space-y-2">
-        <div
-          :for={{row, idx} <- Enum.with_index(@rows)}
-          class="grid grid-cols-[minmax(0,1fr)_minmax(0,1fr)_auto] items-center gap-2"
-        >
-          <input
-            type="text"
-            name={"mcp_endpoint[#{@collection}_rows][#{idx}][key]"}
-            value={row["key"]}
-            placeholder={@key_placeholder}
-            class="w-full font-mono text-[0.82rem] text-black border border-black/10 rounded-lg h-10 px-3 bg-[#fafafa]"
-          />
-          <.secret_input
-            id={"mcp-#{@collection}-#{idx}"}
-            name={"mcp_endpoint[#{@collection}_rows][#{idx}][value]"}
-            value={row["value"]}
-            input_class="w-full font-mono text-[0.82rem] text-black border border-black/10 rounded-lg h-10 px-3 pr-10 bg-[#fafafa]"
-            button_class="absolute right-3 top-1/2 -translate-y-1/2 text-black/30 hover:text-black/60"
-            wrapper_class="relative"
-          />
-          <.mcp_row_delete_button collection={@collection} index={idx} />
-        </div>
-      </div>
-    </div>
-    """
-  end
-
-  attr :collection, :string, required: true
-  attr :index, :integer, required: true
-
-  defp mcp_row_delete_button(assigns) do
-    ~H"""
-    <button
-      type="button"
-      phx-click="remove_mcp_row"
-      phx-value-collection={@collection}
-      phx-value-index={@index}
-      aria-label="Remove row"
-      title="Remove row"
-      class="inline-flex h-10 w-10 items-center justify-center rounded-lg border border-red-200 text-red-500 hover:bg-red-50 hover:text-red-600"
-    >
-      <.icon name="hero-trash" class="h-4 w-4" />
-    </button>
-    """
-  end
-
-  # ── AI Credentials Panel ───────────────────────────────────────────────
-
-  attr :credentials, :list, required: true
-  attr :grants_modal, :boolean, required: true
-  attr :selected_credential, :any, required: true
-  attr :selected_grants, :list, required: true
-  attr :selected_connect_refresh_schedule, :map, required: true
-  attr :credential_modal, :boolean, required: true
-  attr :credential_form, :any, required: true
-  attr :credential_changeset, :any, required: true
-  attr :credential_errors, :list, required: true
-
-  defp connect_credentials_panel(assigns) do
-    ~H"""
-    <div class="bg-white rounded-2xl border border-black/[0.06] shadow-sm overflow-hidden">
-      <div class="px-8 py-5 border-b border-black/[0.06] bg-[#fafafa]">
-        <h2 class="font-mono text-[0.95rem] font-bold text-black">Auth Credentials</h2>
-        <p class="font-mono text-[0.75rem] text-black/40 mt-0.5">
-          Reusable Data Source and MCP credentials with related grants.
-        </p>
-      </div>
-
-      <div :if={@credentials == []} class="px-8 py-10 text-center">
-        <p class="font-mono text-[0.85rem] text-black/50">No credentials configured yet.</p>
-      </div>
-
-      <div :if={@credentials != []} class="divide-y divide-black/[0.06]">
-        <div
-          :for={credential <- @credentials}
-          class="px-8 py-4 flex items-center justify-between gap-4"
-        >
-          <div>
-            <p class="font-mono text-[0.82rem] font-semibold text-black">{credential.name}</p>
-            <p class="font-mono text-[0.7rem] text-black/50 mt-0.5">
-              {credential.provider} · {credential.auth_kind}
-            </p>
-          </div>
-          <div class="flex items-center gap-2">
-            <button
-              type="button"
-              phx-click="open_connect_grants"
-              phx-value-id={credential.id}
-              class="font-mono text-[0.72rem] px-3 py-1.5 rounded-lg border border-black/10 text-black/70 hover:bg-black/[0.04]"
-            >
-              View grants
-            </button>
-            <button
-              type="button"
-              phx-click="edit_connect_credential"
-              phx-value-id={credential.id}
-              class="font-mono text-[0.72rem] px-3 py-1.5 rounded-lg border border-black/10 text-black/70 hover:bg-black/[0.04]"
-            >
-              Edit
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <BOModal.form_dialog
-      :if={@grants_modal}
-      id="connect-grants-modal"
-      cancel_event="close_connect_grants_modal"
-      title={
-        if @selected_credential,
-          do: "Grants — #{@selected_credential.name}",
-          else: "Grants"
-      }
-      max_width_class="max-w-3xl"
-    >
-      <div :if={@selected_grants == []} class="py-8 text-center">
-        <p class="font-mono text-[0.8rem] text-black/50">No grants for this credential.</p>
-      </div>
-
-      <div :if={@selected_grants != []} class="divide-y divide-black/[0.06]">
-        <div :for={grant <- @selected_grants} class="py-3 flex items-start justify-between gap-4">
-          <div class="min-w-0">
-            <div class="flex items-center gap-2">
-              <span class={[
-                "h-2 w-2 rounded-full",
-                grant_expiry_dot_class(grant)
-              ]} />
-              <p class="font-mono text-[0.78rem] font-semibold text-black">
-                {grant.resource_type}:{grant.resource_id}
-              </p>
-            </div>
-            <p class="font-mono text-[0.68rem] text-black/55 mt-0.5">
-              owner={grant.owner_type}:{grant.owner_id || "nil"} · status={grant.status}
-            </p>
-            <p class="font-mono text-[0.68rem] text-black/45 mt-0.5">
-              next refresh: {next_refresh_label(@selected_connect_refresh_schedule, grant)}
-            </p>
-            <p class="font-mono text-[0.68rem] text-black/45 mt-0.5">
-              scopes: {format_grant_scopes(grant.scopes)}
-            </p>
-          </div>
-          <div class="flex items-center gap-2">
-            <p class="font-mono text-[0.68rem] text-black/45">
-              expires: {if grant.expires_at, do: format_grant_datetime(grant.expires_at), else: "none"}
-            </p>
-            <button
-              :if={grant.refresh_token}
-              type="button"
-              phx-click="trigger_connect_grant_refresh"
-              phx-value-id={grant.id}
-              class="font-mono text-[0.68rem] px-2.5 py-1 rounded-lg border border-black/10 text-black/70 hover:bg-black/[0.04]"
-            >
-              Refresh now
-            </button>
-            <button
-              type="button"
-              phx-click="delete_connect_grant"
-              phx-value-id={grant.id}
-              class="font-mono text-[0.68rem] px-2.5 py-1 rounded-lg border border-red-200 text-red-600 hover:bg-red-50"
-            >
-              Erase
-            </button>
-          </div>
-        </div>
-      </div>
-    </BOModal.form_dialog>
-
-    <BOModal.form_dialog
-      :if={@credential_modal}
-      id="edit-connect-credential-modal"
-      cancel_event="close_connect_credential_modal"
-      title="Edit Credential"
-      max_width_class="max-w-lg"
-    >
-      <ConnectCredentialForm.credential_form
-        :if={@credential_form}
-        form={@credential_form}
-        changeset={@credential_changeset}
-        errors={@credential_errors}
-        submit_event="save_connect_credential"
-        change_event="validate_connect_credential"
-        cancel_event="close_connect_credential_modal"
-        id_prefix="edit-connect-credential"
-        submit_label="Save"
-      />
-    </BOModal.form_dialog>
-    """
-  end
-
   defp reload_selected_connect_grants(socket, nil) do
     socket
     |> assign(:selected_connect_grants, [])
@@ -2898,335 +1584,11 @@ defmodule ZaqWeb.Live.BO.System.SystemConfigLive do
   end
 
   defp reload_selected_connect_grants(socket, credential) do
-    grants = Connect.list_grants(credential_id: credential.id)
-    refresh_schedule = Connect.next_refresh_jobs_for_grants(grants)
+    grants = engine_connect_list_grants(credential.id)
+    refresh_schedule = engine_connect_next_refresh_jobs_for_grants(grants)
 
     socket
     |> assign(:selected_connect_grants, grants)
     |> assign(:selected_connect_refresh_schedule, refresh_schedule)
-  end
-
-  defp grant_expiry_dot_class(grant) do
-    cond do
-      is_nil(grant.expires_at) -> "bg-black/25"
-      DateTime.compare(grant.expires_at, DateTime.utc_now()) == :lt -> "bg-red-500"
-      true -> "bg-emerald-500"
-    end
-  end
-
-  defp next_refresh_label(refresh_schedule, grant) do
-    if grant.refresh_token do
-      case Map.get(refresh_schedule, grant.id) do
-        %DateTime{} = dt -> format_grant_datetime(dt)
-        _ -> "none"
-      end
-    else
-      "n/a"
-    end
-  end
-
-  defp format_grant_datetime(%DateTime{} = dt), do: Calendar.strftime(dt, "%Y-%m-%d %H:%M:%SZ")
-
-  defp format_grant_scopes(scopes) when is_list(scopes) do
-    case Enum.reject(scopes, &(&1 in [nil, ""])) do
-      [] -> "none"
-      values -> Enum.join(values, ", ")
-    end
-  end
-
-  defp format_grant_scopes(_), do: "none"
-
-  attr :credentials, :list, required: true
-  attr :form, :any, required: true
-  attr :modal, :boolean, required: true
-  attr :delete_confirm_modal, :boolean, default: false
-  attr :action, :atom, required: true
-
-  defp ai_credentials_panel(assigns) do
-    ~H"""
-    <div class="bg-white rounded-2xl border border-black/[0.06] shadow-sm overflow-hidden">
-      <div class="px-8 py-5 border-b border-black/[0.06] bg-[#fafafa] flex items-center justify-between">
-        <div>
-          <h2 class="font-mono text-[0.95rem] font-bold text-black">AI Credentials</h2>
-          <p class="font-mono text-[0.75rem] text-black/40 mt-0.5">
-            Reusable AI provider credentials used by LLM, Embedding, and Image to Text.
-          </p>
-        </div>
-        <button
-          type="button"
-          phx-click="new_ai_credential"
-          class="font-mono text-[0.75rem] font-bold px-4 py-2 rounded-lg bg-[#03b6d4] text-white hover:bg-[#029ab3] transition-all"
-        >
-          + New credential
-        </button>
-      </div>
-
-      <div :if={@credentials == []} class="px-8 py-10 text-center">
-        <p class="font-mono text-[0.85rem] text-black/50">No AI credentials configured yet.</p>
-      </div>
-
-      <div :if={@credentials != []} class="divide-y divide-black/[0.06]">
-        <button
-          :for={credential <- @credentials}
-          type="button"
-          phx-click="edit_ai_credential"
-          phx-value-id={credential.id}
-          class="w-full text-left px-8 py-4 hover:bg-black/[0.02] transition-all"
-        >
-          <div class="flex items-center justify-between gap-4">
-            <div>
-              <p class="font-mono text-[0.82rem] font-semibold text-black">{credential.name}</p>
-              <p class="font-mono text-[0.7rem] text-black/50 mt-0.5">
-                {credential.provider}
-              </p>
-              <p :if={credential.description} class="font-mono text-[0.7rem] text-black/35 mt-0.5">
-                {credential.description}
-              </p>
-            </div>
-            <span class={[
-              "font-mono text-[0.64rem] px-2 py-1 rounded border",
-              if(credential.sovereign,
-                do: "text-emerald-700 bg-emerald-50 border-emerald-200",
-                else: "text-black/60 bg-black/[0.03] border-black/10"
-              )
-            ]}>
-              {if credential.sovereign, do: "Sovereign", else: "Non-sovereign"}
-            </span>
-          </div>
-        </button>
-      </div>
-    </div>
-
-    <div
-      :if={@modal}
-      class="fixed inset-0 z-50 flex items-center justify-center bg-black/30 backdrop-blur-sm"
-    >
-      <div class="bg-white rounded-2xl shadow-xl border border-black/[0.08] w-full max-w-2xl mx-4 p-6">
-        <h3 class="font-mono text-[0.95rem] font-bold text-black mb-4">
-          {if @action == :edit, do: "Edit AI Credential", else: "New AI Credential"}
-        </h3>
-
-        <.form
-          id="ai-credential-form"
-          for={@form}
-          phx-change="validate_ai_credential"
-          phx-submit="save_ai_credential"
-          class="space-y-4"
-        >
-          <p
-            :for={{msg, opts} <- Keyword.get_values(@form.errors, :base)}
-            class="font-mono text-[0.72rem] text-red-500 bg-red-50 border border-red-100 rounded-xl px-3 py-2"
-          >
-            {translate_error({msg, opts})}
-          </p>
-
-          <div class="grid grid-cols-2 gap-4">
-            <div>
-              <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-                Name
-              </label>
-              <input
-                type="text"
-                name="ai_credential[name]"
-                value={@form[:name].value}
-                class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa]"
-              />
-              <p
-                :for={{msg, opts} <- @form[:name].errors}
-                class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-              >
-                {translate_error({msg, opts})}
-              </p>
-            </div>
-            <div>
-              <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-                Provider
-              </label>
-              <.searchable_select
-                id="ai-credential-provider-select"
-                name="ai_credential[provider]"
-                value={@form[:provider].value}
-                options={provider_options(fn _ -> true end)}
-                placeholder="Search providers..."
-                empty_label="Select a provider..."
-              />
-              <p
-                :for={{msg, opts} <- @form[:provider].errors}
-                class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-              >
-                {translate_error({msg, opts})}
-              </p>
-            </div>
-          </div>
-
-          <div>
-            <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-              Endpoint URL
-            </label>
-            <input
-              type="text"
-              name="ai_credential[endpoint]"
-              value={@form[:endpoint].value}
-              class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 bg-[#fafafa]"
-            />
-            <p
-              :for={{msg, opts} <- @form[:endpoint].errors}
-              class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-            >
-              {translate_error({msg, opts})}
-            </p>
-          </div>
-
-          <div>
-            <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-              API Key
-            </label>
-            <div class="relative">
-              <input
-                type="text"
-                id="ai-credential-api-key-input"
-                name="ai_credential[api_key]"
-                value={@form[:api_key].value}
-                autocomplete="off"
-                style="-webkit-text-security: disc;"
-                class="w-full font-mono text-[0.88rem] text-black border border-black/10 rounded-xl h-11 px-4 pr-10 bg-[#fafafa]"
-              />
-              <button
-                type="button"
-                id="ai-credential-api-key-show"
-                phx-click={
-                  JS.remove_attribute("style", to: "#ai-credential-api-key-input")
-                  |> JS.add_class("hidden", to: "#ai-credential-api-key-show")
-                  |> JS.remove_class("hidden", to: "#ai-credential-api-key-hide")
-                }
-                class="absolute inset-y-0 right-0 flex items-center px-3 text-black/30 hover:text-black/60 transition-colors"
-              >
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke-width="1.5"
-                  stroke="currentColor"
-                  class="w-4 h-4"
-                >
-                  <path
-                    stroke-linecap="round"
-                    stroke-linejoin="round"
-                    d="M2.036 12.322a1.012 1.012 0 0 1 0-.639C3.423 7.51 7.36 4.5 12 4.5c4.638 0 8.573 3.007 9.963 7.178.07.207.07.431 0 .639C20.577 16.49 16.64 19.5 12 19.5c-4.638 0-8.573-3.007-9.963-7.178Z"
-                  /><path
-                    stroke-linecap="round"
-                    stroke-linejoin="round"
-                    d="M15 12a3 3 0 1 1-6 0 3 3 0 0 1 6 0Z"
-                  />
-                </svg>
-              </button>
-              <button
-                type="button"
-                id="ai-credential-api-key-hide"
-                phx-click={
-                  JS.set_attribute({"style", "-webkit-text-security: disc;"},
-                    to: "#ai-credential-api-key-input"
-                  )
-                  |> JS.remove_class("hidden", to: "#ai-credential-api-key-show")
-                  |> JS.add_class("hidden", to: "#ai-credential-api-key-hide")
-                }
-                class="hidden absolute inset-y-0 right-0 flex items-center px-3 text-black/30 hover:text-black/60 transition-colors"
-              >
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke-width="1.5"
-                  stroke="currentColor"
-                  class="w-4 h-4"
-                >
-                  <path
-                    stroke-linecap="round"
-                    stroke-linejoin="round"
-                    d="M3.98 8.223A10.477 10.477 0 0 0 1.934 12C3.226 16.338 7.244 19.5 12 19.5c.993 0 1.953-.138 2.863-.395M6.228 6.228A10.451 10.451 0 0 1 12 4.5c4.756 0 8.773 3.162 10.065 7.498a10.522 10.522 0 0 1-4.293 5.774M6.228 6.228 3 3m3.228 3.228 3.65 3.65m7.894 7.894L21 21m-3.228-3.228-3.65-3.65m0 0a3 3 0 1 0-4.243-4.243m4.242 4.242L9.88 9.88"
-                  />
-                </svg>
-              </button>
-            </div>
-            <p
-              :for={{msg, opts} <- @form[:api_key].errors}
-              class="font-mono text-[0.72rem] text-red-500 mt-1.5"
-            >
-              {translate_error({msg, opts})}
-            </p>
-          </div>
-
-          <div>
-            <label class="flex items-center gap-3 cursor-pointer">
-              <input
-                type="hidden"
-                name="ai_credential[sovereign]"
-                value="false"
-              />
-              <input
-                type="checkbox"
-                name="ai_credential[sovereign]"
-                value="true"
-                checked={@form[:sovereign].value in [true, "true"]}
-                class="sr-only peer"
-              />
-              <div class="w-11 h-6 bg-black/10 peer-checked:bg-[#03b6d4] rounded-full transition-colors after:content-[''] after:absolute after:top-[2px] after:left-[2px] after:bg-white after:rounded-full after:h-5 after:w-5 after:transition-all peer-checked:after:translate-x-5 after:shadow-sm relative">
-              </div>
-              <span class="font-mono text-[0.78rem] text-black/70">Sovereign credential</span>
-            </label>
-          </div>
-
-          <div>
-            <label class="font-mono text-[0.7rem] font-semibold text-black/60 uppercase tracking-wider block mb-2">
-              Description
-            </label>
-            <textarea
-              name="ai_credential[description]"
-              rows="3"
-              class="w-full font-mono text-[0.84rem] text-black border border-black/10 rounded-xl px-4 py-3 bg-[#fafafa]"
-            >{@form[:description].value}</textarea>
-          </div>
-
-          <div class="flex items-center justify-between gap-3 pt-2">
-            <button
-              :if={@action == :edit}
-              type="button"
-              phx-click="open_delete_ai_credential_confirm"
-              class="font-mono text-[0.8rem] px-4 py-2 rounded-lg border border-red-200 text-red-600 hover:bg-red-50"
-            >
-              Delete credential
-            </button>
-
-            <div class="ml-auto flex items-center gap-3">
-              <button
-                type="button"
-                phx-click="close_ai_credential_modal"
-                class="font-mono text-[0.8rem] px-4 py-2 rounded-lg border border-black/10 text-black/60 hover:bg-black/[0.04]"
-              >
-                Cancel
-              </button>
-              <button
-                type="submit"
-                class="font-mono text-[0.8rem] font-bold px-4 py-2 rounded-lg bg-[#03b6d4] text-white hover:bg-[#029ab3]"
-              >
-                Save credential
-              </button>
-            </div>
-          </div>
-        </.form>
-
-        <ZaqWeb.Components.BOModal.confirm_dialog
-          :if={@delete_confirm_modal}
-          id="ai-credential-delete-confirm"
-          cancel_event="cancel_delete_ai_credential"
-          confirm_event="confirm_delete_ai_credential"
-          title="Delete AI Credential?"
-          message="This action removes the credential. Deletion is blocked if the credential is currently in use."
-          confirm_label="Delete"
-          cancel_label="Cancel"
-        />
-      </div>
-    </div>
-    """
   end
 end
